change the data type to handle unix epoch time

+++ migrations/20220624190914_add_fields_needed_for_a_process_lock.sql
@@ -1,5 +1,5 @@
 -- Add migration script here
 ALTER TABLE file_scan
     ADD COLUMN being_worked bool NOT NULL DEFAULT false,
-    ADD COLUMN work_started int;
+    ADD COLUMN work_started bigint;
 CREATE INDEX status_being_worked_idx ON file_scan (being_worked, work_started, status);
